Jakofos 18th May Gryllus bimaculatus (Regulated; unwanted) Lachnus tatakaensis (Regulated; unwanted) Pieris brassicae (Regulated; unwanted) Rhyzobius lophanthae (Regulated) Crassula natans (Controlled; unwanted)
I have added this information to the "species_variables.csv" but I changed the column name from "mentioned_in_EU_quarantine_species" to "mentioned_in_quarantine_species_regulations". Is this okay?

For the methodology it should be clear that this includes EU regulations, the National Pest Plant Accord (NPPA) of NZ and the Pest Register for New Zealand importers.